@@ -69,7 +69,7 @@
       }
     ],
   "require": {
-    "php": ">=7.3.0",
+    "php": ">=7.4.0",
     "ext-bcmath": "*",
     "ext-curl": "*",
     "ext-fileinfo": "*",
@@ -88,7 +88,7 @@
     "doctrine/dbal": "2.*",
     "fideloper/proxy": "4.*",
     "jc5/google2fa-laravel": "2.0.4",
-    "laravel/framework": "^6.0",
+    "laravel/framework": "^7.0",
     "laravel/passport": "8.*",
     "laravelcollective/html": "6.*",
     "league/commonmark": "1.*",
@@ -105,7 +105,7 @@
     "fzaninotto/faker": "1.*",
     "johnkary/phpunit-speedtrap": "^3.1",
     "mockery/mockery": "1.*",
-    "phpunit/phpunit": "8.*",
+    "phpunit/phpunit": "^8.5",
     "psalm/plugin-laravel": "^1.1",
     "roave/security-advisories": "dev-master",
     "vimeo/psalm": "^3.10"
