@@ -146,34 +146,34 @@ return [
     'api_version'                  => '1.1.0',
     'db_version'                   => 13,
     'maxUploadSize'                => 15242880,
-    'send_error_message'           => env('SEND_ERROR_MESSAGE', true),
-    'site_owner'                   => env('SITE_OWNER', ''),
-    'send_registration_mail'       => env('SEND_REGISTRATION_MAIL', true),
-    'demo_username'                => env('DEMO_USERNAME', ''),
-    'demo_password'                => env('DEMO_PASSWORD', ''),
-    'is_sandstorm'                 => env('IS_SANDSTORM', 'unknown'),
-    'bunq_use_sandbox'             => env('BUNQ_USE_SANDBOX', false),
-    'fixer_api_key'                => env('FIXER_API_KEY', ''),
-    'mapbox_api_key'               => env('MAPBOX_API_KEY', ''),
-    'trusted_proxies'              => env('TRUSTED_PROXIES', ''),
-    'search_result_limit'          => env('SEARCH_RESULT_LIMIT', 50),
-    'send_report_journals'         => envNonEmpty('SEND_REPORT_JOURNALS', true),
-    'tracker_site_id'              => env('TRACKER_SITE_ID', ''),
-    'tracker_url'                  => env('TRACKER_URL', ''),
-    'disable_frame_header'         => env('DISABLE_FRAME_HEADER', false),
-    'disable_csp_header'           => env('DISABLE_CSP_HEADER', false),
-    'login_provider'               => envNonEmpty('LOGIN_PROVIDER', 'eloquent'),
-    'cer_provider'                 => envNonEmpty('CER_PROVIDER', 'fixer'),
-    'update_endpoint'              => 'https://version.firefly-iii.org/index.json',
-    'send_telemetry'               => env('SEND_TELEMETRY', false),
-    'telemetry_endpoint'           => 'https://telemetry.firefly-iii.org',
-    'update_minimum_age'           => 6,
-    'default_location'             => [
+    'send_error_message'      => env('SEND_ERROR_MESSAGE', true),
+    'site_owner'              => env('SITE_OWNER', ''),
+    'send_registration_mail'  => env('SEND_REGISTRATION_MAIL', true),
+    'demo_username'           => env('DEMO_USERNAME', ''),
+    'demo_password'           => env('DEMO_PASSWORD', ''),
+    'is_sandstorm'            => env('IS_SANDSTORM', 'unknown'),
+    'fixer_api_key'           => env('FIXER_API_KEY', ''),
+    'mapbox_api_key'          => env('MAPBOX_API_KEY', ''),
+    'trusted_proxies'         => env('TRUSTED_PROXIES', ''),
+    'search_result_limit'     => env('SEARCH_RESULT_LIMIT', 50),
+    'send_report_journals'    => envNonEmpty('SEND_REPORT_JOURNALS', true),
+    'tracker_site_id'         => env('TRACKER_SITE_ID', ''),
+    'tracker_url'             => env('TRACKER_URL', ''),
+    'disable_frame_header'    => env('DISABLE_FRAME_HEADER', false),
+    'disable_csp_header'      => env('DISABLE_CSP_HEADER', false),
+    'login_provider'          => envNonEmpty('LOGIN_PROVIDER', 'eloquent'),
+    'cer_provider'            => envNonEmpty('CER_PROVIDER', 'fixer'),
+    'update_endpoint'         => 'https://version.firefly-iii.org/index.json',
+    'send_telemetry'          => env('SEND_TELEMETRY', false),
+    'telemetry_endpoint'      => 'https://telemetry.firefly-iii.org',
+    'layout'                  => env('FIREFLY_III_LAYOUT', 'v1'),
+    'update_minimum_age'      => 6,
+    'default_location'        => [
         'longitude'  => env('MAP_DEFAULT_LONG', '5.916667'),
         'latitude'   => env('MAP_DEFAULT_LAT', '51.983333'),
         'zoom_level' => env('MAP_DEFAULT_ZOOM', '6'),
     ],
-    'valid_attachment_models'      => [
+    'valid_attachment_models' => [
         Account::class,
         Bill::class,
         Budget::class,
@@ -183,7 +183,7 @@ return [
         Transaction::class,
         TransactionJournal::class,
     ],
-    'allowedMimes'                 => [
+    'allowedMimes'            => [
         /* plain files */
         'text/plain',
 
